Identified and fixed an issue.
I noticed that no forces were being calculated for particles on any thread - This was because each thread was partitioned to do particles 0 to particle 0
Further examination showed that the input pList had size 0, and this was because the threadManager was being init'd before any particles were deployed!

To fix the issue, moved the init of the threadManager to just before the simulation is to begin.