Changed threads to be handled using FSMs.
Each thread will no longer be re-initialised twice per time step - Instead each thread is init'd once when setting up the threadManager, and then the state of the thread is switched using a trivial FSM implementation.
The thread status (what the FSM works off) is mutex'd to correct for an error found before.