Added waiting bool and attempted to fix segfaults
Added a waiting bool and made it the primary way of getting a thread's status publically.
This takes the pressure off the status variable (which is mutex'd) while the main thread is waiting for a workload to be done.
Obviously, the waiting variable is also mutex'd.

Attempted to fix the segfaults by adding in a killPointers() function, similar to the person class.
This may have fixed it, but the code now has another error which I am going to try working on now